flailing

+++ brightboard.cpp
@@ -19,8 +19,8 @@ namespace brightboard {
 	 * @param bits bits per write eg:8
 	 * @param mode mode number eg:3
 	 */
-	//% blockId=brightboard_setDotStarSPIMode block="set bits %bits| set mode %mode"
-	void setDotStarSPIMode(int8_t bits, int8_t mode) {
+	//% blockId=brightboard_setDotStarSPIMode block="set bits %bits| and mode %mode"
+	void setDotStarSPIMode(int bits, int mode) {
 		//spiFormat(8,3);
 		pins::spiFormat(8,0);
 	}
