+++ shims.d.ts
@@ -12,8 +12,8 @@ declare namespace brightboard {
      * @param bits bits per write eg:8
      * @param mode mode number eg:3
      */
-    //% blockId=brightboard_setDotStarSPIMode block="set bits %bits| set mode %mode" shim=brightboard::setDotStarSPIMode
-    function setDotStarSPIMode(bits: int8, mode: int8): void;
+    //% blockId=brightboard_setDotStarSPIMode block="set bits %bits| and mode %mode" shim=brightboard::setDotStarSPIMode
+    function setDotStarSPIMode(bits: int32, mode: int32): void;
 
     /**
      * Send data down the SPI bus
